% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Generator of conversion functions}
\usage{
convert(list1, list2)
}
\arguments{
\item{list1}{A list of items which can be converted.}

\item{list2}{A list of items to be converted two. Both \code{list1} and \code{list2} must
be of the same length as \code{list1}.}
}
\value{
A function which can convert objects in \code{list1} to objects in \code{list2}.
}
\description{
\code{convert} turns a function to convert an item present in \code{list1} to
the same index position in \code{list2}. For example, if \code{list1} contains
("New York", "Paris", "Tokyo") and \code{list2} contains ("United States", "France",
"Japan"), then the function produced will convert city names to country names.
}
