% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{find_kma_parent}
\alias{find_kma_parent}
\title{Takes in a KMA code and returns the parent region of that KMA}
\usage{
find_kma_parent(x)
}
\arguments{
\item{x}{a KMA code in string format to be converted}
}
\value{
the KMA code's parent region
}
\description{
For example, 'TX_DAL' becomes "South Central"
}
