% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{find_region_children}
\alias{find_region_children}
\title{Takes in a Region and returns a string vector of child KMAs}
\usage{
find_region_children(x)
}
\arguments{
\item{x}{a Region in string format to be converted}
}
\value{
a list of KMA codes in string format
}
\description{
For example, "California" returns "CA_FRS" "CA_LAX" "CA_ONT" "CA_SDI" "CA_SFR" "CA_STK" "OR_MED"
}
