% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{kma_to_readable}
\alias{kma_to_readable}
\title{Converts KMA codes into human readable form}
\usage{
kma_to_readable(x)
}
\arguments{
\item{x}{a KMA code in string format to be converted}
}
\value{
a human readable city name
}
\description{
For example, converts "TX_DAL" to "Dallas, TX"
}
