% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{market_areas}
\alias{market_areas}
\title{149 Key Market Area codes and their attributes}
\format{A data frame with 149 rows and 5 variables
\describe{
  \item{city}{US or Canadian City most representative of KMA}
  \item{state_prov}{US State or Canadian province abbreviation}
  \item{kma}{The KMA code}
  \item{full_name}{A properly formatted city and state abbreviation}
  \item{zips}{a list of 3 digit zip codes within the KMA}
}}
\source{
DAT Solutions
}
\usage{
market_areas
}
\description{
149 Key Market Area codes and their attributes
}
\keyword{datasets}
