% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{readable_equip}
\alias{readable_equip}
\title{Convert EQUIPMENT_CATEGORY string into human readable format}
\usage{
readable_equip(x)
}
\arguments{
\item{x}{an equipment type string ('V', 'F', or 'R')}
}
\value{
a human readable string representing that equipment type
}
\description{
Convert EQUIPMENT_CATEGORY string into human readable format
}
