% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{readable_rate_type}
\alias{readable_rate_type}
\title{Convert RATE_TYPE string into human readable format}
\usage{
readable_rate_type(x)
}
\arguments{
\item{x}{a rate type in integer format (1 or 2)}
}
\value{
a human readable string representing that rate type
}
\description{
Convert RATE_TYPE string into human readable format
}
