% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmr_helper.R
\name{readable_use_case}
\alias{readable_use_case}
\title{Produces a string that describes the use case of a time series}
\usage{
readable_use_case(days_back = NULL, rate_type = NULL, equip_type = NULL)
}
\arguments{
\item{days_back}{the days_back parameter of the lane}

\item{rate_type}{the rate_type parameter of the lane}

\item{equip_type}{the equipment_category of the lane}
}
\value{
a string that describes the use case in human readable language
}
\description{
Produces a string that describes the use case of a time series
}
